changed some functions implementation path:   Backend/controllers/admin/adminManagment.js 	changed some functions implementation path:   Backend/controllers/admin/adminPage.js 	bug fix:   Backend/models/estate.js 	changed some functions implementation path:   Backend/routes/admin/adminPage.js 	new file:   Backend/utilities/speciallFunctions.js
