+++ Backend/models/estate.js
@@ -223,7 +223,7 @@ const estateSchema = new mongoose.Schema({
 		type: String,
 		set: (a) => (a === '' ? undefined : a),
 	},
-	costumer_price: {
+	customer_price: {
 		type: String,
 		set: (a) => (a === '' ? undefined : a),
 	},
