+++ Backend/routes/admin/adminPage.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const adminPage_Controller = require('../../controllers/admin/adminPage');
+const regular_controllers = require('../../utilities/speciallFunctions');
 // const adminAuthController = require('../../controllers/admin/adminAuth');
 
 // 2023/05/8 >> changed the logic of written code to 'mounting the routes'
@@ -13,7 +14,7 @@ router
 
 router
 	.route('/estates/generateMint')
-	.post(adminPage_Controller.toLowerCase, adminPage_Controller.generateMint);
+	.post(adminPage_Controller.toLowerCase, regular_controllers.generateMint);
 
 router
 	.route('/estates/:estateId')
